add gallery. tweak styles
