+++ components/layout/index.js
@@ -2,20 +2,28 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import PageHead from '../page-head';
 import MainNavigation from '../main-navigation';
+import FadeIn from '../fade-in';
 import './style.less';
 import '../../pages/main.less';
 
 const Layout = ({
   children,
   pageName,
-  pageTitle
+  pageTitle,
+  headerText
 }) => {
   return (
     <div className={["page-wrapper", pageName].join(' ')}>
       <PageHead pageTitle={pageTitle} />
       <MainNavigation />
-      <main>
-        <div className="main-wrapper">
+      <main className="main-wrapper">
+        <div className="main-container">
+          <FadeIn className="fade-in-up">
+            <h1 className='home__header'>
+              {headerText}
+            </h1>
+            <div className='sexy-line short-line' />
+          </FadeIn>
           {children}
         </div>
       </main>
@@ -26,7 +34,8 @@ const Layout = ({
 export default Layout;
 
 Layout.propTypes = {
-  children: PropTypes.array,
+  children: PropTypes.object,
   pageName: PropTypes.string,
-  pageTitle: PropTypes.string
+  pageTitle: PropTypes.string,
+  headerText: PropTypes.string
 }
