+++ components/main-navigation/style.less
@@ -94,9 +94,9 @@ nav .sexy-line {
     display: inline-block;
     z-index: @exosphere;
     position: absolute;
-    left: 10px;
+    left: 15px;
     height: 36px;
-    padding-top: 20px;
+    padding-top: 25px;
   }
   
   .burg {
