+++ pages/faq/style.less
@@ -1,0 +1,3 @@
+.card {
+  margin-bottom: 30px;
+}
