+++ pages/main.less
@@ -101,6 +101,7 @@ h3,
 .h3 {
   font-size: 1.33333333rem;
   font-weight: 400;
+  margin-bottom: 10px;
   text-transform: uppercase;
 }
 
@@ -136,7 +137,7 @@ p {
   margin-bottom: 1rem;
 }
 
-main {
+.main-wrapper {
   @media @for-phone {
     padding-left: 15px;
     padding-right: 15px;
@@ -146,10 +147,9 @@ main {
   margin-left: auto;
   margin-right: auto;
   padding: 50px 80px 80px;
-  max-width: 1000px
+  max-width: 1000px;
 }
-
-.main-wrapper {
+.main-container {
   @media @for-phone {
     padding-left: 30px;
     padding-right: 30px;
@@ -158,17 +158,16 @@ main {
   width: 100%;
   height: 100%;
   padding: 40px;
+  padding-bottom: 70px;
   box-shadow: 1px 2px 3px rgba(0,0,0,.5);
   position: relative;
 }
 
 section {
-  @media @for-phone {
-    padding-top: 1.5rem;
-    padding-bottom: 1.5rem;
+  @media @for-desktop-up {
+    padding-bottom: 20px;
   }
-  padding-top: 2rem;
-  padding-bottom: 2rem;
+  padding-bottom: 15px;
 }
 
 .subheader {
@@ -195,14 +194,15 @@ section {
 
   &.short-line {
     @media @for-desktop-up {
-      margin-top: 40px;
+      margin-top: 20px;
       margin-bottom: 40px;
     }
-    margin-top: 20px;
+    margin-top: 10px;
     margin-bottom: 20px;
     width: 20%;
   }
 }
+
 /***************************/
 /**** Animation Styles *****/
 /***************************/
@@ -237,7 +237,9 @@ section {
 
 /* set initial values for animations that only occur once upon scroll */
 .fade-in-up,
-.fade-in {
+.fade-in-up--delay1,
+.fade-in,
+.fade-in--delay1 {
   opacity: 0;
   visibility: hidden;
 }
@@ -246,22 +248,22 @@ section {
   transform: translate3d(0, 64px, 0);
 }
 
-// .fade-in-up.active {
-//   animation: .6s .3s fade-in-up ease-out both;
-//   visibility: visible;
-// }
+.fade-in-up.active {
+  animation: .6s .3s fade-in-up ease-out both;
+  visibility: visible;
+}
 
 // .fade-in.active {
 //     animation: .6s .3s fade-in ease-out both;
 //     visibility: visible;
 // }
 
-// .fade-in-up.active--delay1 {
-//   animation: .6s .5s fade-in-up ease-out both;
-//   visibility: visible;
-// }
+.fade-in-up--delay1.active {
+  animation: .6s .5s fade-in-up ease-out both;
+  visibility: visible;
+}
 
-// .fade-in.active--delay1 {
+// .fade-in--delay1.active {
 //   animation: .6s .5s fade-in ease-out both;
 //   visibility: visible;
 // }
