--- pages/photos/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Layout from '../../components/layout'
-
-const Photos = () => (
-  <Layout
-    pageName="photos"
-    pageTitle="Photos"
-  >
-    <h1>Photos</h1>
-    <p className="subheader">Here are a few photos of us and some of our favorite memories.</p>
-  </Layout>
-)
-
-export default Photos
