--- pages/registry/index.tsx
@@ -1,30 +1,0 @@
-import React from 'react'
-import Layout from '../../components/layout';
-
-const Registry = () => (
-  <Layout
-    pageTitle="Registry"
-    pageName="registry"
-  >
-    <h1>Registry</h1>
-    <section>
-      <p>Hello friends and family!</p>
-      <p>
-        Thank you for visiting our registry page. Please know that the greatest gift you could
-        give us would be your attendance to our wedding and sharing time with us!
-      </p>
-      <p>
-        We live in a cozy little house in East Austin and don't have space for too many things.
-        Although we have a few traditional gifts on our registry and some household upgrades,
-        we decided to make our honeymoon, kick starting our new married life, and a few
-        charities the focal point of our registry. If you'd like, please help send us along
-        on a memorable trip to Croatia by contributin to some of our honeymoon cash gift experiences.
-      </p>
-      <p>
-        Thank you so much for keeping us in your minds and hearts. We're so grateful for all of you!
-      </p>
-    </section>
-  </Layout>
-)
-
-export default Registry
